requirements.txt
langchain
langchain-community
langchain-experimental

fastapi
python-dotenv
streamlit
uvicorn
pydantic
httpx
requests

langchain-google-community
langchain-tavily
langchain-groq
groq
langchain-openai

pyyaml

langgraph
langsmith

-e .

